Really passing all the test

+++ src/assemble/assemble.h
@@ -24,7 +24,7 @@ void assembleMultiplyInstruction(FILE* fpOutput, struct Token*);
 void assembleSingleDataInstruction(FILE* fpOutput, struct Token*);
 void assembleBranchInstruction(FILE* fpOutput, struct Token*);
 
-uint16_t getOperand2Immediate(uint32_t operand2Val);
+uint16_t getOperand2Immediate(long operand2Val);
 uint16_t getOperand2ShiftRegister(uint32_t operand2Val);
 struct SymbolTable* initializeInstructionCodeTable();
